+++ VisualPinball.Engine/VPT/Gate/GateData.cs
@@ -96,7 +96,7 @@ namespace VisualPinball.Engine.VPT.Gate
 
 		public override void Write(BinaryWriter writer, HashWriter hashWriter)
 		{
-			writer.Write(ItemType.Gate);
+			writer.Write((int)ItemType.Gate);
 			Write(writer, Attributes, hashWriter);
 			WriteEnd(writer, hashWriter);
 		}
